+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * Fixes app title for ios
 * Updates the marker for the Solar Car, so it looks better
 * Updates the team details page with information about every team member.
-* Full leaderboard page is added to the app
+* Implements fill leaderboard page
 
 ## 0.0.2